application/requirements.txt
streamlit
pandas
matplotlib
seaborn
cassandra-driver
pytz
plotly